User can navigate between taps
